# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_restaurant.AvailabilityResponse:
    properties:
      business_date:
        example: "2026-10-10"
        type: string
      closed:
        description: วันปิดประจำสัปดาห์ → slots ว่าง
        type: boolean
      closes_at:
        type: string
      opens_at:
        type: string
      seats:
        type: integer
      slots:
        items:
          $ref: '#/definitions/internal_restaurant.SlotResponse'
        type: array
    type: object
  internal_restaurant.ImageRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  internal_restaurant.ImageResponse:
    properties:
      id:
        type: string
      sort_order:
        type: integer
      url:
        type: string
    type: object
  internal_restaurant.ListItemResponse:
    properties:
      address:
        type: string
      cancel_before_minutes:
        type: integer
      close_time:
        example: "02:00"
        type: string
      closed_weekdays:
        description: วันปิดประจำสัปดาห์ของวันทำการ (0 = อาทิตย์ … 6 = เสาร์)
        items:
          type: integer
        type: array
      created_at:
        type: string
      cuisine:
        type: string
      description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/internal_restaurant.ImageResponse'
        type: array
      name:
        type: string
      open_24h:
        type: boolean
      open_time:
        example: "18:00"
        type: string
      overnight:
        description: ปิดหลังเที่ยงคืน → หน้าเว็บต้องแสดงป้าย "(เช้าวันที่ n)"
        type: boolean
      owner_id:
        type: string
      rating:
        $ref: '#/definitions/internal_restaurant.RatingResponse'
      seats:
        type: integer
      slots:
        items:
          $ref: '#/definitions/internal_restaurant.SlotResponse'
        type: array
    type: object
  internal_restaurant.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_restaurant.ListItemResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  internal_restaurant.NextAvailableResponse:
    properties:
      business_date:
        type: string
      slots:
        items:
          $ref: '#/definitions/internal_restaurant.SlotResponse'
        type: array
    type: object
  internal_restaurant.RatingResponse:
    properties:
      average:
        description: null = ยังไม่มีรีวิว
        type: number
      count:
        type: integer
    type: object
  internal_restaurant.RestaurantRequest:
    properties:
      address:
        maxLength: 300
        type: string
      cancel_before_minutes:
        example: 30
        maximum: 1440
        minimum: 30
        type: integer
      close_time:
        example: "22:00"
        type: string
      closed_weekdays:
        description: 0 = อาทิตย์ … 6 = เสาร์
        example:
        - 1
        items:
          type: integer
        type: array
      cuisine:
        example: อาหารไทย
        maxLength: 60
        type: string
      description:
        maxLength: 2000
        type: string
      image_urls:
        items:
          type: string
        maxItems: 10
        type: array
      name:
        example: ครัวบ้านสวน
        maxLength: 120
        type: string
      open_time:
        example: "11:00"
        type: string
      seats:
        example: 10
        maximum: 1000
        type: integer
    required:
    - address
    - cancel_before_minutes
    - close_time
    - name
    - open_time
    - seats
    type: object
  internal_restaurant.RestaurantResponse:
    properties:
      address:
        type: string
      cancel_before_minutes:
        type: integer
      close_time:
        example: "02:00"
        type: string
      closed_weekdays:
        description: วันปิดประจำสัปดาห์ของวันทำการ (0 = อาทิตย์ … 6 = เสาร์)
        items:
          type: integer
        type: array
      created_at:
        type: string
      cuisine:
        type: string
      description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/internal_restaurant.ImageResponse'
        type: array
      name:
        type: string
      open_24h:
        type: boolean
      open_time:
        example: "18:00"
        type: string
      overnight:
        description: ปิดหลังเที่ยงคืน → หน้าเว็บต้องแสดงป้าย "(เช้าวันที่ n)"
        type: boolean
      owner_id:
        type: string
      rating:
        $ref: '#/definitions/internal_restaurant.RatingResponse'
      seats:
        type: integer
    type: object
  internal_restaurant.SlotResponse:
    properties:
      available:
        type: integer
      closed:
        type: boolean
      end_at:
        type: string
      start_at:
        type: string
    type: object
...
  jongyoung_internal_httputil.ErrorBody:
    properties:
      code:
        example: NOT_ENOUGH_SEATS
        type: string
      details: {}
      message:
        example: ที่นั่งไม่พอ
        type: string
    type: object
  jongyoung_internal_httputil.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/jongyoung_internal_httputil.ErrorBody'
    type: object
host: api.jongyoung.localhost
info:
  contact: {}
  description: ระบบจองโต๊ะร้านอาหาร + รีวิว — เวลาทั้งหมดเป็น timestamp เต็ม (RFC
    3339) ส่วน date = วันทำการ
  title: jongyoung API
  version: "1.0"
paths:
...
  /restaurants:
    get:
      parameters:
      - description: ชื่อร้าน/เมนู
        in: query
        name: q
        type: string
      - description: ประเภทอาหาร
        in: query
        name: cuisine
        type: string
      - description: rating | reviews | newest
        in: query
        name: sort
        type: string
      - description: หน้า
        in: query
        name: page
        type: integer
      - description: จำนวนต่อหน้า (สูงสุด 50)
        in: query
        name: limit
        type: integer
      - description: วันทำการ YYYY-MM-DD
        in: query
        name: date
        type: string
      - description: เวลาประมาณ HH:MM
        in: query
        name: time
        type: string
      - description: จำนวนคน
        in: query
        name: party_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_restaurant.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      summary: ค้นหาร้าน (+ ปุ่มเวลาว่าง 5 ช่วงเมื่อส่ง date, time, party_size มาครบ)
      tags:
      - restaurants
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลร้าน (ต้องมี image_urls อย่างน้อย 1)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_restaurant.RestaurantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_restaurant.RestaurantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: สร้างร้าน (ผู้สร้างเป็นเจ้าของ)
      tags:
      - restaurants
  /restaurants/{id}:
    delete:
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ลบร้าน (soft delete) และยกเลิกการจองที่ยังไม่เริ่มทั้งหมด
      tags:
      - restaurants
    get:
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_restaurant.RestaurantResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      summary: รายละเอียดร้าน
      tags:
      - restaurants
    put:
      consumes:
      - application/json
      description: ลดที่นั่งต่ำกว่าคนสูงสุดของการจองที่จะถึง หรือย่นเวลาจนการจองตกนอกเวลา
        → 409
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลร้าน (ไม่ใช้ image_urls)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_restaurant.RestaurantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_restaurant.RestaurantResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: แก้ไขร้าน (เฉพาะเจ้าของ)
      tags:
      - restaurants
  /restaurants/{id}/availability:
    get:
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      - description: วันทำการ YYYY-MM-DD (รอบที่เริ่มเปิดในวันนั้น)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_restaurant.AvailabilityResponse'
      summary: ที่ว่างทุกช่วง 30 นาทีของรอบวันทำการ
      tags:
      - restaurants
...
  /restaurants/{id}/images:
    post:
      consumes:
      - application/json
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      - description: URL รูป
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_restaurant.ImageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_restaurant.ImageResponse'
      security:
      - BearerAuth: []
      summary: เพิ่มรูปร้าน (URL)
      tags:
      - restaurants
  /restaurants/{id}/images/{imageId}:
    delete:
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      - description: image id
        in: path
        name: imageId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ลบรูปร้าน (ห้ามลบรูปสุดท้าย)
      tags:
      - restaurants
  /restaurants/{id}/next-available:
    get:
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      - description: วันทำการที่ค้น YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      - description: HH:MM
        in: query
        name: time
        required: true
        type: string
      - description: จำนวนคน
        in: query
        name: party_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_restaurant.NextAvailableResponse'
      summary: วันทำการถัดไป (ไม่เกิน 14 วัน) ที่มีช่วงรอบเวลาที่ค้นว่างพอ — ไม่เจอคืน
        null
      tags:
      - restaurants
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: พิมพ์ "Bearer " ตามด้วย access token จาก Keycloak
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"